requirements.txt
mangum
fastapi[all]
fastapi-sso
python-jose[cryptography]